feat: Add Carousel component for displaying proposal screenshots
This commit adds a new Carousel component to the project. The Carousel component is used to display proposal screenshots in a slideshow format. It accepts an array of images as a prop and renders them in a carousel layout. This component enhances the user experience by providing an interactive way to view multiple screenshots of a proposal.